requirements.txt
sentence-transformers
pinecone-client
langchain
flask
ctransformers
langchain-community
pypdf
langchain-pinecone
python-dotenv

-e .